Mechanism to chose different atom pairs in the FES
Added mechanism to select which atoms should be included in the FES
reconstruction by the chapeau decomposition of the pair potential
function. An index number is associated to each atom with a value of
-1, 0, 1 or 2. I call this number the "mask" number. Then, only those
pairs of atoms that have a sum of their mask numbers equal to 0 are
included in the chapeau object.  Therefore, the atoms with a mask number
of 1 will match only with those atoms with a mask number of -1.  On the
other hand, atoms with a mask number of 0 only match atoms that also
have a 0. A mask number of 2 means that the atom is always excluded. The
mask label is controlled per residue from the namd input file through
the variable "pairresnames". For instance:

set pairresnames      "SOD CLA"

Label the SOD atoms with -1 and CLA atoms with 1. Al the others will
have assigned a mask number of 2. On the other hand, the entry

set pairresnames      "SOD SOD"

gives a maks number of 0 to the SOD atoms, while the others will have a
mask of 2.  If pairresnames is not set, all the atoms will get a mask
value of 0, making old namd configuration files compatible with this new
feature.