New example: NaCl
- Folder examples/NaCl is added to test the develop of the program.
- The mask control allow to select the free energy term to recover. For
example, if the auxiliary variables are the cartesian coordinates of the
NaCl ions, then the FES associated with the TAMD acceleration will be
something like:

$$
FES=\sum_{i>j}^{Na} g_{NaNa}(r_{ij})
   +\sum_{i>j}^{Cl} g_{ClCl}(r_{ij})
   +\sum_i^{Cl}\sum_j^{Na} g_{NaCl}(r_{ij})
$$

However, while all the ions are accelerated, might be the
parameterization trough the chapeau functions is desired only for one of
the pair potentials (i.e. $g_{NaCl}$, $g_{NaCl}$ or $g_{NaCl}$). This
can be specified in the execution trough the mask used (see previous
commit). In the next version I will vectorize the chapeau object to
compute the three pair potentials in the same simulation.