Mayor changes: TCL modules, replica restart, testing
The procedures in cfacv.tcl were rearranged into different tcl files, according
to the chapeau, restraint, centers and cvs classes. The clacforces procedure
and addgroup namd commands are place also in a separate file. The remaining
code is sourced in the input file only to initialize the objects and variables
of the program. In this way, it is possible to initialize the code and then
modify the objects before the first namd run command. This gives more
flexibility to the namd input file, allowing to modify the program setup (e.g.
    restart the chapeaus) before run. At the same time some procedures included
in the code can now be moved to the input, which increase portability.

The restart system allows now two different modes for loading the chapeau
state: load the full object (with A and b matrices) or only the lambda
parameters. Then the variables runs_noevolvech and runs_nosolvech in the
replica exchange scripts allow to control the evolution and solution of the
chapeau objects. This changes follows from the idea of using two REOTFP: the
equilibration and the production stage. The first one is used to separate the
replicas from the initial configuration and the second one to compute the free
energy. The first steps of the equilibration stage occurs without exchanges,
since we don't have any free energy to use in the acceptance rule. During the
first steps of the production stage the exchanges are controlled by the free
energy obtained in the equilibration and the new free energy is computed in the
background (without being used). Then, after some steps, the old free energy is
discarded and the new is used for the acceptance rule, which is refined during
the rest of the simulation.

Under the folder examples I added new simulations with butane that can serve
for testing future versions of the code. In this sense, I include also a test
script to run and get some plots. The references plots are also included for
comparision.

FIX in catbinsp output format