New feature: CV boundaries. BUG: COGZ, COGX and related
In order to restrain a CV to a certain domain, boundary forces can be added
directly to a CV. Note that this is not the same as setting boundaries to a
restraint, since the natural motion of the CV is not tied to a virtual
particle. To implement this major changes to the CV input were introduced.

A bug in some CV calc procedures was fix: The mixing of index in the ds->R[]
matrix and the cv->ind[] array.