Output system for cvs. New CVs: RMSD and LINE.
The output system for the cv object was added and works in the same way that
output system for the restraints.

RMSD is an example of a CV that require more parameters that common CV
structures, i.e. a reference for the centers position. While an inheritance
hierarchy could be more adequate for this situation, C does not have
inheritance/polymorphism capabilities. An hybrid C++/C code would be the final
solution but this should be done with a careful checking of the C++/swig
interface capabilities. For now I added a pointer (ref) to the CV
structure that will be null for the common cv structures and only be
used when the CV is a RMSD.

In the case of the LINE CV, another pointer was added (ref2) since two
references are needed. I think that this can be avoided since ref2 can
be stored in the cv->gr, but I let this work for later.

Other changes:

- Some structures names and some types names are simplified, avoiding
the redundant "struct" suffix.

- read_centersPDB transform to parse_pdb and moved to cvs.tcl. parse_pdb
can read a pdb file and construct a pdb array with all the pdb data for
further processing. The rest of center.tcl is not needed/used and thus
the file is removed.