Periodicity in stiffness matrix
Periodicity was not included properly in the previous commits. Now the data is
equally accumulated in the boundaries chapeaus.  Still, I am not sure if there
is something wrong now...  the resulting FES does not seems to be fully
periodic, since after translation I still can see a line in the boundary (see
`examples/OTFP2D_periodic/reference/job1_ch0_fes.png` and translate with gimp
or similar). Not sure if this can/should be avoided or if there is a bug in the
preiodic treatment included in this commit.

On the other hand, atom index is now independent of the pdb number (like in
namd), allowing larger indexes without taking care of the column width of the
pdb index field.