FIX: REOTFP with lapack library.
Work on REOTFP-2D.

The bug can bee seen with `git log --follow -p -- chapeau.c`. The
relevant part of this file is:

    commit 424467e218d920cca0117b2bc1036221ccd52aa2
    Author: S. Alexis Paz <alexis.paz@gmail.com>
    Date:   Fri May 13 17:30:44 2016 -0400

    diff --git a/src/chapeau.c b/src/chapeau.c
    index edb55ab..c4343f0 100644

The mistake is the change of:

    -  //diagonal
    -  memcpy(word1, &str[size1], 13 ); size1+=13;
    -  err=sscanf(word1,"%13le",&aux); if(!err) {fprintf(stderr,"CFACV/C) Error 2114 on read %s\n",word1);}
    -  ch->Afull[i][i]=aux;
    -
    -  //off diagonal
    -  memcpy(word1, &str[size1], 13 ); size1+=13;
    -  err=sscanf(word1,"%13le",&aux); if(!err) {fprintf(stderr,"CFACV/C) Error 2115 on read %s\n",word1);}
    -  ch->Afull[i][i-1]=aux;
    -  ch->Afull[i-1][i]=aux;

To:

    +  // Reading the partial A (simetric and tridiagonal)
    +  memcpy(word1, &str[size1], 13 ); size1+=13;
    +  err=sscanf(word1,"%13le",&aux); if(!err) {fprintf(stderr,"CFACV/C) Error 1112 on read %s\n",word1);}
    +  ch->Afull[i][j]+=aux;

The bug is the += since it should be only = since the replica should
acept the full statistic from the core replica that performed the
addition and not add it again. When OTFP was extended to 2D, gsl
libraries was changed to lapack and the A matrix change the way to be
stored at memory. At that moment the bug was introduced.